rename cookie to sessionStorage
